requirements.txt
# IAM
bk_iam

# OT
protobuf==3.19.5
opentelemetry-api==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-exporter-otlp==1.11.1
